+++ NEWS.md
@@ -1,3 +1,10 @@
+# quickcode 0.2
+
+- [x] New functions
+- [x] Add 'seed' argument to functions using randomness
+- [x] Update to clean and refresh functions
+
+
 # quickcode 0.1
 
 - [x] Initial release